+++ packages/odyssey-react-mui/src/Badge.tsx
@@ -27,7 +27,7 @@ export type BadgeProps = {
   badgeContent: number;
   badgeContentMax?: BadgeContentMax;
   type?: (typeof badgeTypeValues)[number];
-} & HtmlProps;
+} & Pick<HtmlProps, "testId" | "translate">;
 
 const badgeTypeColors = (odysseyTokens: DesignTokens) => ({
   default: {
@@ -89,7 +89,7 @@ const Badge = ({
     const hasNotificationCount = badgeContent && badgeContent > 0;
 
     return hasNotificationCount ? (
-      <Box sx={badgeStyles} data-se={testId} translate={translate}>
+      <Box sx={badgeStyles} testId={testId} translate={translate}>
         {formattedContent}
       </Box>
     ) : null;