+++ packages/odyssey-react-mui/src/Breadcrumbs.tsx
@@ -43,7 +43,7 @@ export type BreadcrumbsProps = {
   children: ReactElement<typeof Breadcrumb>[];
   homeHref?: string;
   maxVisibleItems?: number;
-} & HtmlProps;
+} & Pick<HtmlProps, "testId" | "translate">;
 
 export type BreadcrumbContextType = {
   breadcrumbType: BreadcrumbType;
@@ -101,13 +101,18 @@ const defaultTruncationValue = 5;
 const BreadcrumbList = ({
   children,
   homeHref,
-  maxVisibleItems = defaultTruncationValue,
+  maxVisibleItems: maxVisibleItemsProp = defaultTruncationValue,
   testId,
   translate,
 }: BreadcrumbsProps) => {
   const { t } = useTranslation();
 
   const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);
+
+  const maxVisibleItems = useMemo(
+    () => Math.min(Math.max(maxVisibleItemsProp, 0), children.length),
+    [maxVisibleItemsProp, children],
+  );
 
   const breadcrumbSections = useMemo(() => {
     if (children.length <= maxVisibleItems) {
@@ -155,7 +160,7 @@ const BreadcrumbList = ({
         </BreadcrumbContext.Provider>
       ))}
 
-      {breadcrumbSections.insideMenu && (
+      {breadcrumbSections.insideMenu.length > 0 && (
         <>
           <ButtonBase onClick={onMenuButtonClick}>...</ButtonBase>
           <Menu