+++ packages/odyssey-react-mui/src/labs/index.ts
@@ -20,7 +20,8 @@ export * from "./datePickerTheme";
 /** @deprecated Will be removed in a future Odyssey version in lieu of the one shipping with DataTable */
 export * from "./DataTablePagination";
 export * from "./DataFilters";
-
+export * from "./Drawer";
+export * from "./FileUpload";
 export * from "./materialReactTableTypes";
 /** @deprecated Will be removed in a future Odyssey version in lieu of DataTable */
 export * from "./StaticTable";