+++ packages/odyssey-icons/CHANGELOG.md
@@ -3,6 +3,48 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.14.9](https://github.com/okta/odyssey/compare/v1.14.8...v1.14.9) (2024-03-08)
+
+**Note:** Version bump only for package @okta/odyssey-icons
+
+## [1.14.8](https://github.com/okta/odyssey/compare/v1.14.7...v1.14.8) (2024-03-08)
+
+**Note:** Version bump only for package @okta/odyssey-icons
+
+## [1.14.7](https://github.com/okta/odyssey/compare/v1.14.6...v1.14.7) (2024-03-08)
+
+**Note:** Version bump only for package @okta/odyssey-icons
+
+## [1.14.6](https://github.com/okta/odyssey/compare/v1.14.5...v1.14.6) (2024-03-07)
+
+**Note:** Version bump only for package @okta/odyssey-icons
+
+## [1.14.5](https://github.com/okta/odyssey/compare/v1.14.4...v1.14.5) (2024-03-01)
+
+**Note:** Version bump only for package @okta/odyssey-icons
+
+## [1.14.4](https://github.com/okta/odyssey/compare/v1.14.3...v1.14.4) (2024-03-01)
+
+**Note:** Version bump only for package @okta/odyssey-icons
+
+## [1.14.3](https://github.com/okta/odyssey/compare/v1.14.2...v1.14.3) (2024-02-29)
+
+**Note:** Version bump only for package @okta/odyssey-icons
+
+## [1.14.2](https://github.com/okta/odyssey/compare/v1.14.1...v1.14.2) (2024-02-27)
+
+**Note:** Version bump only for package @okta/odyssey-icons
+
+## [1.14.1](https://github.com/okta/odyssey/compare/v1.14.0...v1.14.1) (2024-02-26)
+
+**Note:** Version bump only for package @okta/odyssey-icons
+
+## [1.14.0](https://github.com/okta/odyssey/compare/v1.13.12...v1.14.0) (2024-02-26)
+
+### Bug Fixes
+
+- Upgrade every package ([#2136](https://github.com/okta/odyssey/issues/2136)) ([89b6534](https://github.com/okta/odyssey/commit/89b65342755a30d0e947f5ffb75e6bbe00268442))
+
 ## [1.13.12](https://github.com/okta/odyssey/compare/v1.13.11...v1.13.12) (2024-02-16)
 
 **Note:** Version bump only for package @okta/odyssey-icons